Add execution of version.sh script before build in snapcraft.yaml
This commit modifies the snapcraft.yaml file to run the version.sh script
before the build phase. This is achieved by adding the script execution in
the 'override-pull' hook.

The version.sh script is responsible for updating the version number, which is
subsequently displayed in the About dialog of the built program.

Without this change, the About dialog would only display the build time.