Use new craftctl and CRAFT_* vars
This commit updates the snapcraft build script to address a
warning related to the transition from "snapcraftctl" to
"craftctl", and the introduction of the new "CRAFT_*"
environment variables.

All instances of "snapcraftctl" have been replaced with
"craftctl", and all "SNAPCRAFT_*" environment variables used
in the script have been replaced with their "CRAFT_*"
equivalents.

This change is necessary to adapt to the latest version of
Snapcraft, which is shifting towards using "craftctl" and
"CRAFT_*" environment variables for better support of newer
Snapcraft bases such as core22, as well as improved build
tools and methodologies.

More details on this transition can be found in the official
Snapcraft documentation and the core22 migration guide.

With this update, the build process should no longer
generate the warning, and the script should be better
aligned with the latest Snapcraft practices.